Changelog: -Replantear el sistema de la conexion a la base de datos, he hecho rollback todos los cambios relacionados con ella no se puede tocar sin destruir nada, hay que hacer que pueda ser multiusuario y que haya multiples conexiones a la vez. - Tocados algunos imports, voy a intentar no hacer servir de los archivos resources si no es necesario. - LoginForm.py ahora si estas en modo debug te setea el user como Admin - Cambiado el typo del campo estadio de Paciente en entities.py

+++ database/entities.py
@@ -428,7 +428,7 @@ class Pacient(Entity):
         columns.add_column("dni", "text", "PRIMARY KEY")
         columns.add_column("nombre", "text")
         columns.add_column("apellidos", "text")
-        columns.add_column("estadio", "integer")
+        columns.add_column("estadio", "float")
         columns.add_column("nacimiento", "date")
         columns.add_column("notas", "text")
         columns.add_column("telefono", "integer")
